feat: new router structure
Working towards making this less trello dependent.
