+++ lib/db.js
@@ -179,6 +179,11 @@ class Storage {
    */
   async getByID (id) {
     const res = await this.adapter.query('SELECT * FROM media WHERE id = $1', [id])
+    if (res.rows.length === 0) {
+      const err = new Error('ID not found')
+      err.code = 'ERRNOTFOUND'
+      throw err
+    }
     if (res.rows.length !== 1) throw new Error('Found multiple rows on single op')
 
     const row = res.rows[0]
