+++ lib/trello.js
@@ -14,8 +14,6 @@ const logger = require('pino')({
   name: path.basename(__filename)
 })
 
-let webhookDef = null
-
 /**
  * Create a trello webhook proccessor if needed.
  * Then setup the webhook listener.
@@ -35,11 +33,6 @@ const init = async (auth, opts) => {
 
   logger.info('trello board id is', board.id)
   opts.board = board.id
-
-  if (!webhookDef) {
-    logger.debug('created webhook')
-    webhookDef = require('./webhook')(opts.event, opts.tracer)
-  }
 
   // bail out of registering the webhook, but still run it
   // i.e for local development
